小改注释；点击采集后，只能就不能更改采集时间了。

+++ customchart.h
@@ -19,7 +19,7 @@ public:
 
 public slots:
     void addYPoint(double y);
-    void addPoint(int x, double y);
+    void addPoint(double x, double y);
 
     void addVLine(double x);
     void addHLine(double y);
